Rename project and folder StellarWallet.Test to  StellarWallet.UnitTest
