+++ StellarWallet.UnitTest/StellarWallet.UnitTest.csproj
